skybet/requirements.txt
Flask==1.1.1
redis==3.3.7
requests==2.22
rq==1.1.0
selenium==3.141.0
python-dateutil==2.8.0